* forgot to add the sam/print functions

+++ src/util/fmap_sam.c
@@ -41,7 +41,7 @@ fmap_sam_print_unmapped(fmap_file_t *fp, fmap_seq_t *seq)
 inline void
 fmap_sam_print_mapped(fmap_file_t *fp, fmap_seq_t *seq, fmap_refseq_t *refseq,
                     uint8_t strand, uint32_t seqid, uint32_t pos, 
-                    int32_t mapq, uint32_t *cigar, int32_t n_cigar,
+                    uint8_t mapq, uint32_t *cigar, int32_t n_cigar,
                     const char *format, ...)
 {
   va_list ap;
